System.FastLogger

Documentation

data Logger

Holds the state for a logger.

timestampedLogEntry :: ByteString -> IO ByteString

Prepares a log message with the time prepended.

combinedLogEntry

Arguments

:: ByteString
remote host

-> Maybe ByteString
remote user

-> ByteString
request line (up to you to ensure there are no quotes in here)

-> Int
status code

-> Maybe Int64
num bytes sent

-> Maybe ByteString
referer (up to you to ensure there are no quotes in here)

-> ByteString
user agent (up to you to ensure there are no quotes in here)

-> IO ByteString

Prepares a log message in "combined" format.

newLogger :: FilePath -> IO Logger

Creates a new logger, logging to the given file. If the file argument is "-", then log to stdout; if it's "stderr" then we log to stderr, otherwise we log to a regular file in append mode. The file is closed and re-opened every 15 minutes to facilitate external log rotation.

logMsg :: Logger -> ByteString -> IO ()

Sends out a log message verbatim with a newline appended. Note: if you want a fancy log message you'll have to format it yourself (or use combinedLogEntry).

stopLogger :: Logger -> IO ()

Kills a logger thread, causing any unwritten contents to be flushed out to disk